crash course by codeholic completed

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate();
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderBy('created_at', 'desc')
+            ->paginate();
         // dd($notes);
         return view('note.index', ['notes' => $notes]);
     }
@@ -34,11 +37,11 @@ class NoteController extends Controller
             'note' => ['required', 'string'],
         ]);
 
-        // $data['user_id'] = auth()->user()->id;
-        $data['user_id'] = 1;
+        $data['user_id'] = request()->user()->id;
+        // $data['user_id'] = 1;
         $note = Note::create($data);
 
-        return to_route('note.show', $note) -> with('message', 'Note created successfully');
+        return to_route('note.show', $note)->with('message', 'Note created successfully');
     }
 
     /**
@@ -46,6 +49,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
@@ -54,6 +60,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -62,13 +71,17 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+
         $data = $request->validate([
             'note' => ['required', 'string'],
         ]);
 
         $note->update($data);
 
-        return to_route('note.show', $note) -> with('message', 'Note updated successfully');
+        return to_route('note.show', $note)->with('message', 'Note updated successfully');
     }
 
     /**
@@ -76,9 +89,12 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
 
         $note->delete();
 
-        return to_route('note.index') -> with('message', 'Note deleted successfully');
+        return to_route('note.index')->with('message', 'Note deleted successfully');
     }
 }
